backend/requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.4.0
django-rest-framework==0.1.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
npm==0.1.1
oauthlib==3.1.0
optional-django==0.1.0
pep8==1.7.1
Pillow==7.2.0
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.3
pypiwin32==223
python3-openid==3.2.0
pytz==2020.1
pywin32==228
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
websocket-client==0.57.0
wrapt==1.12.1
yarn==0.0.9
